save legal agreement at time of creation

+++ E-EYE-O_Interfaces/src/com/jtbdevelopment/e_eye_o/helpandlegal/PrivacyPolicy.java
@@ -1,7 +1,5 @@
 package com.jtbdevelopment.e_eye_o.helpandlegal;
 
-import java.math.BigDecimal;
-
 /**
  * Date: 5/7/13
  * Time: 10:02 AM
@@ -14,5 +12,5 @@ public interface PrivacyPolicy {
 
     String getText();
 
-    BigDecimal getVersion();
+    int getVersion();
 }
